Loads more stuff for py2exe...
Getting there slowly. Having discovered that distutils (and setuptools
and py2exe and basically everything to do with packaging in Python) is
completely fucking broken when it comes to unicode, it's time to remove
unicode_literals from setup.py.

Having done that, and then discovered that the previous magic of
include_package_data relied on CVS and SVN specifically (read as:
doesn't work with git/hg/anything decent), it's time to rejig everything
to work without it.

Rhetorical question of the moment: is anything to do with Python
packaging written and tested and not utterly fucking retarded?!
(disclaimer: this statement is born of hours of frustration and ignores
several rather cool things like package resources and entry points which
work rather well)