Major renaming
Turns out py2exe doesn't like constructing exes for things in different
directories. Furthermore, turns out that if a script is based directly
in a package (i.e. in the __init__ module within that package) it'll
wind up with the name __init__.exe (*headdesk*). So, the
rastools.rasviewer package is renamed to rastools.windows (implying that
all GUI modules live under it) while rastools.rasviewer becomes a module
within the rastools package...

All for the benefit of a bloody Windows exe...