More make changes...
Added README and LICENSE to the docs in Debian builds, sorted out proper
subdir dependencies for building PNGs in icons and fallback-theme,
removed the pointless package maintainer stuff for RPMs in setup.cfg (it
gets defaulted to Vendor anyway, which is defaulted in turn to Author).
Removed the old .exe build for Windows and added the beginnings of a
rudimentary MSI build. Removed the pointless (and odd) "build" prefix
from the various package targets. And finally, updated the make help
target.