Fixed percentile going crazy when panning
If the range controls were focused, and the percentile controls were set
to values other than 0 and 100 while a zoomed portion of the image was
panned, the percentile controls would go nuts. This was because the
crop_changed handler was manipulating the range controls which were in
turn manipulating the percentile controls. At the next pan, the
percentile would be recalculated, adjusting the value and the whole
thing would loop around again getting madder as the panning continued
until a maxima or minima was reached and the picture showed a single
value.

Now the percentile controls are considered locked during crop
adjustments so value may wander but percentile will not.